% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Catapult_Fn.R
\name{Catapult_CSV_ImportAndConvert}
\alias{Catapult_CSV_ImportAndConvert}
\title{Imports the file from CSV and converts it to the correct format}
\usage{
Catapult_CSV_ImportAndConvert(sFileName)
}
\arguments{
\item{A}{CSV file}
}
\description{
Imports the file from CSV and converts it to the correct format
}
